Separate components on stake card
