+++ src/pages/Staking/StakeCard/BalanceTree.tsx
@@ -1,0 +1,49 @@
+import { FC } from "react"
+import { StakeData } from "../../../types/staking"
+import { Tree, TreeNode } from "../../../components/Tree"
+import { Icon, Tooltip } from "@chakra-ui/react"
+import { InfoIcon } from "@chakra-ui/icons"
+import { BalanceTreeItem } from "./BalanceTreeItem"
+
+export const BalanceTree: FC<{ stake: StakeData }> = ({ stake }) => {
+  return (
+    <Tree>
+      <TreeNode isRoot>
+        <BalanceTreeItem
+          isRoot
+          label={
+            <>
+              Total Staked Balance{" "}
+              <Tooltip
+                label="Staked Balance for Legacy Stakes are cumulated KEEP, NU and T staked tokens displayed in T."
+                fontSize="md"
+                bg="white"
+                color="gray.700"
+                textAlign="center"
+                p="2"
+                offset={[150, 10]}
+                hasArrow
+              >
+                <Icon as={InfoIcon} />
+              </Tooltip>
+            </>
+          }
+          value={stake.totalInTStake}
+        >
+          <TreeNode>
+            <BalanceTreeItem label="Native Stake" value={stake.tStake} />
+            {stake.keepInTStake !== "0" && (
+              <BalanceTreeItem
+                label="KEEP Stake in T"
+                value={stake.keepInTStake}
+              />
+            )}
+            {stake.nuInTStake !== "0" && (
+              <BalanceTreeItem label="NU Stake in T" value={stake.nuInTStake} />
+            )}
+          </TreeNode>
+        </BalanceTreeItem>
+      </TreeNode>
+    </Tree>
+  )
+}
